Dev

See merge request s08-webmobile2-sub2/S08P12A807!386

+++ togetdog/src/main/java/com/ssafy/togetdog/user/model/service/JwtServiceImpl.java
@@ -58,7 +58,7 @@ public class JwtServiceImpl implements JwtService {
 	}
 
 	@Override
-	public <T> String create(long userId, String subject, long expir) {
+	public <T> String create(long userId, String subject, int expir) {
 		String jwt = Jwts.builder()
 				.setHeaderParam("type", "JWT")
 				.setHeaderParam("regDate", System.currentTimeMillis())
@@ -71,7 +71,7 @@ public class JwtServiceImpl implements JwtService {
 	}
 
 	@Override
-	public <T> String create(long userId, String subject, long expir, String role) {
+	public <T> String create(long userId, String subject, int expir, String role) {
 		String jwt = Jwts.builder()
 				.setHeaderParam("type", "JWT")
 				.setHeaderParam("regDate", System.currentTimeMillis())
